requirements.txt
streamlit
gradio
torch
numpy
Pillow